backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
yfinance
pandas
httpx
google-generativeai
python-dotenv 